requirements.txt
django==3.2.6
mysqlclient
django_bootstrap5
django-widget-tweaks
factory_boy
Faker